# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: 華繰ルナ 公開家計簿 API
  version: "1.0.0"
  description: |
    LLM / 外部スクリプト向け。視聴者向け HTML とは別。

    使い方:
    1. GET /api/v1/meta で科目を覚える
    2. GET /api/v1/status で今の数字を読む
    3. 必要なら明細・予定・お知らせ・4コマ・企画・挑戦を取る
    4. 登録するときだけ POST / PATCH

    認証: `Authorization: Bearer <RUNA_API_TOKEN>`
    単位: 円の整数。日付: yyyy-MM-dd。日時: ISO 8601（推奨 Asia/Tokyo）
    NMR の円は評価額。総資産には入れる。損益・自給率・現金には入れない。
servers:
  - url: http://localhost:3000
    description: ローカル
  - url: /
    description: このサイト
security:
  - bearerAuth: []
tags:
  - name: meta
  - name: status
  - name: projects
  - name: announcements
  - name: comics
  - name: ideas
  - name: transactions
  - name: events
paths:
  /api/v1/meta:
    get:
      tags: [meta]
      summary: 科目と種別の辞書。最初に読む
      responses:
        "200":
          description: 辞書
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/status:
    get:
      tags: [status]
      summary: 生存KPI・総資産・NMR円・直近の明細と予定。施策を考える入口
      responses:
        "200":
          description: 現状。assets.total は現金+機材+NMR円。nmr.amount は総NMR、nmr.staked は Stake中。未確定の Payout/Burn は損益に入れない
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/projects:
    get:
      tags: [projects]
      summary: 挑戦一覧。取引・予定・企画の projectId に使う
      responses:
        "200":
          description: "{ projects: Project[] }"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [projects]
      summary: 挑戦作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectWrite"
            example:
              title: Numeraiで稼ぐ
              status: active
              masterNote: null
              overview: 予測を出してNMRを稼ぐ
              links: []
      responses:
        "201":
          description: 作成した挑戦
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/projects/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [projects]
      summary: 挑戦1件
      responses:
        "200":
          description: 挑戦
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [projects]
      summary: 挑戦更新（送った項目だけ）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectPatch"
      responses:
        "200":
          description: 更新後の挑戦
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [projects]
      summary: 挑戦削除。明細・予定・募集案が付いているときは 422
      responses:
        "200":
          description: "{ ok: true, id }"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /api/v1/announcements:
    get:
      tags: [announcements]
      summary: お知らせ一覧。未公開も含む。published=true で公開だけ
      parameters:
        - in: query
          name: published
          schema: { type: string, enum: ["true"] }
        - in: query
          name: category
          schema:
            $ref: "#/components/schemas/AnnouncementCategory"
      responses:
        "200":
          description: "{ announcements: Announcement[] }"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [announcements]
      summary: お知らせ作成。画像は coverUrl（公開URL）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementWrite"
            example:
              title: 配信します
              body: 今夜20時からだよ
              category: stream
              publishedAt: "2026-09-06"
              coverUrl: null
              published: true
      responses:
        "201":
          description: 作成したお知らせ
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/announcements/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [announcements]
      summary: お知らせ1件
      responses:
        "200":
          description: お知らせ
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [announcements]
      summary: お知らせ更新（送った項目だけ）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementPatch"
      responses:
        "200":
          description: 更新後のお知らせ
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [announcements]
      summary: お知らせ削除。予定が付いているときは 422
      responses:
        "200":
          description: "{ ok: true, id }"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /api/v1/comics:
    get:
      tags: [comics]
      summary: 4コマ一覧。未公開も含む。published=true で公開だけ
      parameters:
        - in: query
          name: published
          schema: { type: string, enum: ["true"] }
      responses:
        "200":
          description: "{ comics: ComicStrip[] }"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [comics]
      summary: 4コマ作成。画像は imageUrl（公開URL）が必須
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComicWrite"
            example:
              title: 1話
              imageUrl: https://example.com/yonkoma.png
              published: true
              sortOrder: 0
      responses:
        "201":
          description: 作成した4コマ
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/comics/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [comics]
      summary: 4コマ1件
      responses:
        "200":
          description: 4コマ
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [comics]
      summary: 4コマ更新（送った項目だけ）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComicPatch"
      responses:
        "200":
          description: 更新後の4コマ
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [comics]
      summary: 4コマ削除
      responses:
        "200":
          description: "{ ok: true, id }"
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/ideas:
    get:
      tags: [ideas]
      summary: 企画一覧。status で絞れる
      parameters:
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/IdeaStatus"
      responses:
        "200":
          description: "{ ideas: Idea[] }"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [ideas]
      summary: 企画作成（公開フォームの honeypot は無し）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdeaWrite"
            example:
              displayName: 匿名
              title: 配信で料理してみる
              body: 所持金の範囲で一品作る企画だよ
              status: submitted
              projectId: null
      responses:
        "201":
          description: 作成した企画
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/ideas/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [ideas]
      summary: 企画1件
      responses:
        "200":
          description: 企画
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [ideas]
      summary: 企画更新。ステータスと挑戦の紐づけはここ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdeaPatch"
            example:
              status: adopted
              projectId: clxxx
      responses:
        "200":
          description: 更新後の企画
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [ideas]
      summary: 企画削除
      responses:
        "200":
          description: "{ ok: true, id }"
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/transactions:
    get:
      tags: [transactions]
      summary: 明細一覧
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
          description: 開始日 yyyy-MM-dd。省略時は90日前
        - in: query
          name: to
          schema: { type: string, format: date }
          description: 終了日 yyyy-MM-dd。省略時は今日（日本時間）
        - in: query
          name: type
          schema:
            $ref: "#/components/schemas/TxType"
      responses:
        "200":
          description: "{ transactions: Transaction[] }"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [transactions]
      summary: 明細作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionWrite"
            example:
              date: "2026-08-26"
              type: expense
              category: llm_api
              amount: 184
              title: Claude API
              memo: null
              projectId: null
      responses:
        "201":
          description: 作成した明細
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/Error"
  /api/v1/transactions/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [transactions]
      summary: 明細1件
      responses:
        "200":
          description: 明細
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [transactions]
      summary: 明細更新（送った項目だけ）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionPatch"
      responses:
        "200":
          description: 更新後の明細
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
    delete:
      tags: [transactions]
      summary: 明細削除
      responses:
        "200":
          description: "{ ok: true, id }"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /api/v1/events:
    get:
      tags: [events]
      summary: 予定一覧（期間と重なるもの）
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
          description: 省略時は60日前
        - in: query
          name: to
          schema: { type: string, format: date }
          description: 省略時は今日
      responses:
        "200":
          description: "{ events: Event[] }"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [events]
      summary: 予定作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventWrite"
            example:
              title: 配信
              startAt: "2026-08-28T20:00:00+09:00"
              endAt: "2026-08-28T22:00:00+09:00"
              allDay: false
              kind: stream
              body: null
              linkUrl: null
              projectId: null
              announcementId: null
      responses:
        "201":
          description: 作成した予定
        "400":
          $ref: "#/components/responses/Error"
  /api/v1/events/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [events]
      summary: 予定1件
      responses:
        "200":
          description: 予定
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [events]
      summary: 予定更新（送った項目だけ）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventPatch"
      responses:
        "200":
          description: 更新後の予定
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      tags: [events]
      summary: 予定削除
      responses:
        "200":
          description: "{ ok: true, id }"
        "404":
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string }
  responses:
    Unauthorized:
      description: トークンが無い・違う。未設定のときは 503 TOKEN_UNSET
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Error:
      description: エラー
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    TxType:
      type: string
      enum: [income, expense, loan, repay, capex]
    EventKind:
      type: string
      enum: [stream, release, project, other]
    ProjectStatus:
      type: string
      enum: [planned, active, completed]
    AnnouncementCategory:
      type: string
      enum: [news, stream, other]
    IdeaStatus:
      type: string
      enum: [submitted, reviewing, adopted, in_progress, done]
    ProjectLink:
      type: object
      required: [url]
      properties:
        kind: { type: string, enum: [youtube, note, other] }
        label: { type: string, maxLength: 40 }
        url: { type: string }
    TransactionWrite:
      type: object
      required: [date, type, amount, category, title]
      properties:
        date:
          type: string
          format: date
        type:
          $ref: "#/components/schemas/TxType"
        amount:
          type: integer
          minimum: 1
          description: 円。常に正
        category:
          type: string
          description: |
            income: ads, superchat, merch, affiliate, support, ai_hustle
            expense: llm_api, voice, hosting, tools, other
            loan: master_loan
            repay: master_repay
            capex: equipment
        title:
          type: string
          maxLength: 80
        memo:
          type: [string, "null"]
        projectId:
          type: [string, "null"]
    TransactionPatch:
      type: object
      description: 更新する項目だけ送る
      properties:
        date: { type: string, format: date }
        type:
          $ref: "#/components/schemas/TxType"
        amount: { type: integer, minimum: 1 }
        category: { type: string }
        title: { type: string }
        memo: { type: [string, "null"] }
        projectId: { type: [string, "null"] }
    EventWrite:
      type: object
      required: [title, startAt, endAt, kind]
      properties:
        title: { type: string, maxLength: 80 }
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        allDay: { type: boolean }
        kind:
          $ref: "#/components/schemas/EventKind"
        body: { type: [string, "null"] }
        linkUrl: { type: [string, "null"] }
        projectId: { type: [string, "null"] }
        announcementId: { type: [string, "null"] }
    EventPatch:
      type: object
      description: 更新する項目だけ送る
    ProjectWrite:
      type: object
      required: [title, status]
      properties:
        title: { type: string, maxLength: 80 }
        status:
          $ref: "#/components/schemas/ProjectStatus"
        masterNote: { type: [string, "null"], maxLength: 500 }
        overview: { type: [string, "null"], maxLength: 400 }
        links:
          type: array
          maxItems: 8
          items:
            $ref: "#/components/schemas/ProjectLink"
    ProjectPatch:
      type: object
      description: 更新する項目だけ送る
    AnnouncementWrite:
      type: object
      required: [title, body, category, publishedAt]
      properties:
        title: { type: string, maxLength: 80 }
        body: { type: string, maxLength: 4000 }
        category:
          $ref: "#/components/schemas/AnnouncementCategory"
        publishedAt:
          type: string
          format: date
        coverUrl: { type: [string, "null"] }
        published: { type: boolean }
    AnnouncementPatch:
      type: object
      description: 更新する項目だけ送る
    ComicWrite:
      type: object
      required: [title, imageUrl]
      properties:
        title: { type: string, maxLength: 80 }
        imageUrl:
          type: string
          format: uri
          description: 公開URL。ファイルアップロードは管理画面
        published: { type: boolean }
        sortOrder: { type: integer, minimum: 0, maximum: 9999 }
    ComicPatch:
      type: object
      description: 更新する項目だけ送る
    IdeaWrite:
      type: object
      required: [title, body]
      properties:
        displayName: { type: [string, "null"], maxLength: 40 }
        title: { type: string, maxLength: 80 }
        body: { type: string, minLength: 10, maxLength: 1000 }
        status:
          $ref: "#/components/schemas/IdeaStatus"
        projectId: { type: [string, "null"] }
    IdeaPatch:
      type: object
      description: 更新する項目だけ送る。ステータス更新は status と任意の projectId
      properties:
        displayName: { type: [string, "null"], maxLength: 40 }
        title: { type: string, maxLength: 80 }
        body: { type: string, minLength: 10, maxLength: 1000 }
        status:
          $ref: "#/components/schemas/IdeaStatus"
        projectId: { type: [string, "null"] }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [TOKEN_UNSET, UNAUTHORIZED, VALIDATION, SOLVENCY, CONFLICT, NOT_FOUND, INTERNAL]
            message: { type: string }
            hint: { type: string }